type : refactor 코드 정리
body :
