+++ src/test/java/bridge/ApplicationTest.java
@@ -3,14 +3,17 @@ package bridge;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberInRangeTest;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.util.Lists.newArrayList;
 
 import bridge.main.Application;
 import bridge.model.BridgeMaker;
 import bridge.model.BridgeNumberGenerator;
 import bridge.model.BridgeRandomNumberGenerator;
+import bridge.validator.Validator;
 import camp.nextstep.edu.missionutils.test.NsTest;
 import java.util.List;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class ApplicationTest extends NsTest {
@@ -58,8 +61,25 @@ class ApplicationTest extends NsTest {
         BridgeMaker bridgeMaker = new BridgeMaker(new BridgeRandomNumberGenerator());
         int size = 10;
         bridge = bridgeMaker.makeBridge(size);
-        System.out.println(bridge);
         assertThat(bridge.size()).isEqualTo(size);
+    }
+
+    @Test
+    @DisplayName("플레이어 이동 시 올바른 입력")
+    void correctMove() {
+        assertThatCode(() -> Validator.checkMoveInput("U"))
+                .doesNotThrowAnyException();
+
+        assertThatCode(() -> Validator.checkMoveInput("D"))
+                .doesNotThrowAnyException();
+
+    }
+
+    @Test
+    @DisplayName("다리길이 올바른 입력")
+    void correctBridgeSizeInput() {
+        assertThatCode(() -> Validator.checkBridgeSize(10))
+                .doesNotThrowAnyException();
     }
 
     @Override
